Move a bunch more commands that depended on pkg/cmd/templates
