+++ hack/import-restrictions.json
@@ -412,7 +412,6 @@
       "github.com/openshift/origin/pkg/oc/cli/admin/groups/examples"
     ],
     "allowedImportPackageRoots": [
-      "vendor/github.com/docker/go-units",
       "vendor/github.com/fsnotify/fsnotify",
       "vendor/github.com/fsouza/go-dockerclient",
       "vendor/k8s.io/klog",
@@ -426,7 +425,6 @@
       "vendor/github.com/spf13/pflag",
       "vendor/k8s.io/api",
       "vendor/k8s.io/apimachinery",
-      "vendor/k8s.io/apiserver",
       "vendor/k8s.io/client-go",
       "vendor/k8s.io/cli-runtime",
       "vendor/k8s.io/klog",
@@ -443,7 +441,6 @@
     "allowedImportPackages": [
       "github.com/openshift/origin/pkg/build/apis/build",
       "github.com/openshift/origin/pkg/build/apis/build/install",
-      "github.com/openshift/origin/pkg/cmd/templates",
       "github.com/openshift/origin/pkg/image/apis/image",
       "github.com/openshift/origin/pkg/project/apiserver/registry/projectrequest/delegated",
       "github.com/openshift/origin/pkg/version",
