+++ pkg/oc/cli/cli.go
@@ -29,36 +29,36 @@ import (
 	"github.com/openshift/oc/pkg/cli/expose"
 	"github.com/openshift/oc/pkg/cli/extract"
 	"github.com/openshift/oc/pkg/cli/idle"
+	"github.com/openshift/oc/pkg/cli/image"
 	"github.com/openshift/oc/pkg/cli/kubectlwrappers"
 	"github.com/openshift/oc/pkg/cli/login"
 	"github.com/openshift/oc/pkg/cli/logout"
 	"github.com/openshift/oc/pkg/cli/logs"
 	"github.com/openshift/oc/pkg/cli/observe"
 	"github.com/openshift/oc/pkg/cli/options"
+	"github.com/openshift/oc/pkg/cli/policy"
 	"github.com/openshift/oc/pkg/cli/process"
 	"github.com/openshift/oc/pkg/cli/project"
 	"github.com/openshift/oc/pkg/cli/projects"
 	"github.com/openshift/oc/pkg/cli/recycle"
+	"github.com/openshift/oc/pkg/cli/registry"
 	"github.com/openshift/oc/pkg/cli/requestproject"
 	"github.com/openshift/oc/pkg/cli/rollback"
+	"github.com/openshift/oc/pkg/cli/rollout"
 	"github.com/openshift/oc/pkg/cli/rsh"
 	"github.com/openshift/oc/pkg/cli/secrets"
 	"github.com/openshift/oc/pkg/cli/serviceaccounts"
+	"github.com/openshift/oc/pkg/cli/set"
 	"github.com/openshift/oc/pkg/cli/status"
 	"github.com/openshift/oc/pkg/cli/whoami"
+	"github.com/openshift/oc/pkg/helpers/templates"
 	"github.com/openshift/oc/pkg/helpers/term"
 
-	"github.com/openshift/origin/pkg/cmd/templates"
 	"github.com/openshift/origin/pkg/oc/cli/admin"
-	"github.com/openshift/origin/pkg/oc/cli/image"
 	"github.com/openshift/origin/pkg/oc/cli/importimage"
 	"github.com/openshift/origin/pkg/oc/cli/newapp"
 	"github.com/openshift/origin/pkg/oc/cli/newbuild"
-	"github.com/openshift/origin/pkg/oc/cli/policy"
-	"github.com/openshift/origin/pkg/oc/cli/registry"
-	"github.com/openshift/origin/pkg/oc/cli/rollout"
 	"github.com/openshift/origin/pkg/oc/cli/rsync"
-	"github.com/openshift/origin/pkg/oc/cli/set"
 	"github.com/openshift/origin/pkg/oc/cli/startbuild"
 	"github.com/openshift/origin/pkg/oc/cli/tag"
 )
