+++ staging/src/github.com/openshift/oc/pkg/cli/admin/policy/modify_scc_test.go
