+++ staging/src/github.com/openshift/oc/pkg/cli/admin/policy/remove_from_project.go
