+++ staging/src/github.com/openshift/oc/pkg/cli/policy/policy.go
@@ -5,8 +5,8 @@ import (
 	"k8s.io/cli-runtime/pkg/genericclioptions"
 	kcmdutil "k8s.io/kubernetes/pkg/kubectl/cmd/util"
 
-	adminpolicy "github.com/openshift/origin/pkg/oc/cli/admin/policy"
-	"github.com/openshift/origin/pkg/oc/cli/policy/cani"
+	adminpolicy "github.com/openshift/oc/pkg/cli/admin/policy"
+	"github.com/openshift/oc/pkg/cli/policy/cani"
 )
 
 const PolicyRecommendedName = "policy"
