+++ staging/src/github.com/openshift/oc/pkg/cli/rollout/latest.go
