+++ staging/src/github.com/openshift/oc/pkg/cli/set/deploymenthook.go
