+++ staging/src/github.com/openshift/oc/pkg/cli/set/imagelookup.go
