+++ staging/src/github.com/openshift/oc/pkg/cli/set/probe.go
