+++ staging/src/github.com/openshift/oc/pkg/cli/set/set_test.go
