+++ staging/src/github.com/openshift/oc/pkg/helpers/templates/templater.go
