+++ test/integration/pod_node_constraints_test.go
@@ -11,7 +11,7 @@ import (
 	rbacv1client "k8s.io/client-go/kubernetes/typed/rbac/v1"
 
 	appsclient "github.com/openshift/client-go/apps/clientset/versioned"
-	"github.com/openshift/origin/pkg/oc/cli/admin/policy"
+	"github.com/openshift/oc/pkg/cli/admin/policy"
 	pluginapi "github.com/openshift/origin/pkg/scheduler/admission/apis/podnodeconstraints"
 	testutil "github.com/openshift/origin/test/util"
 	testserver "github.com/openshift/origin/test/util/server"
